src: move JobOutput/Builds/Tests in separate component files
