+++ src/JobStats.js
@@ -1,4 +1,4 @@
-import { useState, useEffect, useCallback } from "react";
+import { useState } from "react";
 
 const Worker = (props) => {
     return (
